weight/tfidfweight.cc: Explicitly cast argument of log() to double, as xlC doesn't manage to chose an overloaded form to call otherwise.
git-svn-id: file:///home/xapian-svn/xapian/trunk@17426 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-core/weight/tfidfweight.cc
@@ -129,7 +129,7 @@ TfIdfWeight::get_wdfn(Xapian::termcount wdf, char c) const
 	    return (wdf * wdf);
 	case 'l':
 	    if (wdf == 0) return 0;
-	    return (1 + log(wdf));
+	    return (1 + log(double(wdf)));
 	default:
 	    AssertEq(c, 'n');
 	    return wdf;
